requirements.txt
# Google / YouTube
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Web dashboard
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Environment
python-dotenv>=1.0.0

# HTTP / utilities
requests>=2.31.0
